created client and server
Created two program, one for the client and one for the server. The
client got two threads, one reading from the server and the other
writing to the server.
The server have likewise one write and one read. The write writes to
everyone while the read only reads from a specific participant.
The server is extending the Mailbox class.
It's running, but I haven't gotten the client to communicate yet.